# docs/api/task-api-spec.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API for managing tasks in a productivity application.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server

paths:
  /tasks:
    get:
      summary: Get a list of tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, in progress, completed]
          description: Filter tasks by status
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high]
          description: Filter tasks by priority
        - in: query
          name: dueDate
          schema:
            type: string
            format: date
          description: Filter tasks by due date
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{id}:
    get:
      summary: Get a single task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    put:
      summary: Update an existing task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Task deleted

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in progress, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TaskInput:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in progress, completed]
          